BACKEND/requirements.txt
Django==5.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2-binary